refactor(DFPlayerMiniFast): optimize and fix style issues
- Remove redundant recStack initialization in begin() — parseFeedback()
  overwrites all fields on arrival
- Fix findChecksum narrowing: move +1 inside static_cast to avoid
  uint16_t→unsigned int promotion on assignment
- Rename volume() parameter to level to avoid shadowing the method name
- Change timeoutTime and threshold parameters from uint32_t to uint16_t
  (max useful timeout fits in 65535 ms)
- Fix constructor/destructor declaration order in header

+++ lib/dfPlayer/src/dfPlayer.cpp
@@ -3,7 +3,7 @@
 
 volatile bool DFPlayer::enablePlay = false;                         // Set value for static variable.
 
-DFPlayer::DFPlayer(RgbLedWrapper& rgbLed, uint8_t rxPin, uint8_t txPin, uint8_t enPin, uint8_t intPin, uint32_t timeout) :
+DFPlayer::DFPlayer(RgbLedWrapper& rgbLed, uint8_t rxPin, uint8_t txPin, uint8_t enPin, uint8_t intPin, uint16_t timeout) :
   rgbLed(rgbLed),
   swSerial(rxPin, txPin),
   rxPin(rxPin),
